refactor(ebpf): split eBPF programs by operation
