+++ plugins/krsi/krsi-ebpf/src/operations.rs
@@ -1,0 +1,4 @@
+pub mod open;
+pub mod connect;
+pub mod socket;
+mod close;
