+++ plugins/krsi/krsi-ebpf/src/operations/open.rs
