+++ plugins/krsi/krsi-ebpf/src/operations/open/maps.rs
