+++ plugins/krsi/krsi-ebpf/src/operations/socket/maps.rs
